+++ src/models/GitHubActionOptions.ts
@@ -9,4 +9,12 @@ export interface GitHubActionOptions {
   showError?: boolean;
   quiet?: boolean;
   includeResults?: DisplayLevel[];
+  debug?: boolean;
+
+  // Useful links
+  showArtifactsLink?: boolean;
+
+  // Azure Storage
+  azureStorageUrl?: string;
+  azureStorageSAS?: string;
 }