+++ src/utils/processResults.ts
@@ -10,7 +10,7 @@ import { getSummaryTitle } from "./getSummaryTitle";
 import { getSummaryDetails } from "./getSummaryDetails";
 import { getTestsPerFile } from "./getTestsPerFile";
 import { getTestHeading } from "./getTestHeading";
-import { DisplayLevel, GitHubActionOptions } from "../models";
+import { BlobService, DisplayLevel, GitHubActionOptions } from "../models";
 
 export const processResults = async (
   suite: Suite | undefined,
@@ -29,6 +29,18 @@ export const processResults = async (
   if (process.env.GITHUB_ACTIONS && suite) {
     const os = process.platform;
     const summary = core.summary;
+
+    let blobService: BlobService | undefined = undefined;
+    if (options.azureStorageSAS && options.azureStorageUrl) {
+      blobService = {};
+      blobService.azure = {};
+      blobService.azure.azureStorageSAS = options.azureStorageSAS;
+      blobService.azure.azureStorageUrl = options.azureStorageUrl;
+    }
+
+    if (options.showArtifactsLink) {
+      summary.addLink("Go to artifacts", "#artifacts");
+    }
 
     const summaryTitle = getSummaryTitle(options.title);
     if (summaryTitle) {
@@ -58,6 +70,7 @@ export const processResults = async (
             !!options.showError,
             options.includeResults as DisplayLevel[],
             options.showAnnotationsInColumn,
+            blobService
           );
 
           if (!content) {
@@ -79,6 +92,7 @@ export const processResults = async (
             !!options.showError,
             options.includeResults as DisplayLevel[],
             options.showAnnotationsInColumn,
+            blobService
           );
 
           if (tableRows.length !== 0) {